\name{.create_andrews_tour}
\alias{.create_andrews_tour}
\title{Andrews Tour Plotting...}
\usage{.create_andrews_tour(data, var_selected, cat_selected, dim_selected,
    tour_type, guided_type, lambda, aps)}
\description{Andrews Tour Plotting
Plots the Andrews Tour}
\keyword{internal}
\author{Bei Huang\email{beihuang@iastate.edu}, Di Cook \email{dicook@iastate.edu}, and Hadley Wickham \email{hadley@rice.edu}}
