\name{.create_face_tour}
\alias{.create_face_tour}
\title{Faces Tour Plotting...}
\usage{.create_face_tour(data, var_selected, nface_selected, cat_selected,
    dim_selected, tour_type, guided_type, lambda, aps)}
\description{Faces Tour Plotting
Plots the Faces Tour}
\keyword{internal}
\author{Bei Huang\email{beihuang@iastate.edu}, Di Cook \email{dicook@iastate.edu}, and Hadley Wickham \email{hadley@rice.edu}}
