\name{.create_image_tour}
\alias{.create_image_tour}
\title{Image Tour Plotting...}
\usage{.create_image_tour(data, tour_type, aps)}
\description{Image Tour Plotting
Plots the Image Tour}
\keyword{internal}
\author{Bei Huang\email{beihuang@iastate.edu} and Di Cook \email{dicook@iastate.edu}}
