\name{.create_mat_tour}
\alias{.create_mat_tour}
\title{Scatmat Tour Plotting...}
\usage{.create_mat_tour(data, var_selected, projdim_selected, tour_type, aps)}
\description{Scatmat Tour Plotting
Plots the Scatmat Tour}
\keyword{internal}
\author{Bei Huang\email{beihuang@iastate.edu}, Di Cook \email{dicook@iastate.edu}, and Hadley Wickham \email{hadley@rice.edu}}
