\name{.create_pcp_tour}
\alias{.create_pcp_tour}
\title{PCP Tour Plotting...}
\usage{.create_pcp_tour(data, var_selected, dim_selected, tour_type, aps)}
\description{PCP Tour Plotting
Plots the PCP Tour}
\keyword{internal}
\author{Bei Huang\email{beihuang@iastate.edu} and Di Cook \email{dicook@iastate.edu}}
