\name{.create_stars_tour}
\alias{.create_stars_tour}
\title{Stars Tour Plotting...}
\usage{.create_stars_tour(data, var_selected, dim_selected, tour_type, aps)}
\description{Stars Tour Plotting
Plots the Stars Tour}
\keyword{internal}
\author{Bei Huang\email{beihuang@iastate.edu}, Di Cook \email{dicook@iastate.edu}, and Hadley Wickham \email{hadley@rice.edu}}
