\name{.create_stereo_tour}
\alias{.create_stereo_tour}
\title{Stereo Tour Plotting...}
\usage{.create_stereo_tour(data, var_selected, tour_type, aps)}
\description{Stereo Tour Plotting
Plots the Stereo Tour}
\keyword{internal}
\author{Bei Huang\email{beihuang@iastate.edu}, Di Cook \email{dicook@iastate.edu}, and Hadley Wickham \email{hadley@rice.edu}}
