\name{.create_xy_tour}
\alias{.create_xy_tour}
\title{Scatterplot Tour Plotting...}
\usage{.create_xy_tour(data, var_selected, cat_selected, axes_location,
    tour_type, guided_type, lambda, angle, aps)}
\description{Scatterplot Tour Plotting
Plots the scatterplot Tour}
\details{Creates tour for gui_xy}
\keyword{internal}
\author{Bei Huang\email{beihuang@iastate.edu}, Di Cook \email{dicook@iastate.edu}, and Hadley Wickham \email{hadley@rice.edu}}
