\name{.interface_andrews}
\alias{.interface_andrews}
\title{Andrews Tour Plotting...}
\usage{.interface_andrews(g5, data, w)}
\description{Andrews Tour Plotting
Plots the Andrews Tour in tab g5}
\keyword{internal}
\author{Bei Huang\email{beihuang@iastate.edu}, Di Cook \email{dicook@iastate.edu}, and Hadley Wickham \email{hadley@rice.edu}}
