\name{.interface_density}
\alias{.interface_density}
\title{Density Tour Plotting...}
\usage{.interface_density(g2, data, w)}
\description{Density Tour Plotting
Plots the Density Tour in tab g2}
\keyword{internal}
\author{Bei Huang\email{beihuang@iastate.edu}, Di Cook \email{dicook@iastate.edu}, and Hadley Wickham \email{hadley@rice.edu}}
