\name{.interface_faces}
\alias{.interface_faces}
\title{Faces Tour Plotting...}
\usage{.interface_faces(g3, data, w)}
\description{Faces Tour Plotting
Plots the Faces Tour in tab g3}
\keyword{internal}
\author{Bei Huang\email{beihuang@iastate.edu}, Di Cook \email{dicook@iastate.edu}, and Hadley Wickham \email{hadley@rice.edu}}
