\name{.interface_pcp}
\alias{.interface_pcp}
\title{PCP Tour Plotting...}
\usage{.interface_pcp(g7, data, w)}
\description{PCP Tour Plotting
Plots the PCP Tour in tab g7}
\keyword{internal}
\author{Bei Huang\email{beihuang@iastate.edu}, Di Cook \email{dicook@iastate.edu}, and Hadley Wickham \email{hadley@rice.edu}}
