\name{.interface_scatmat}
\alias{.interface_scatmat}
\title{Scatmat Tour Plotting...}
\usage{.interface_scatmat(g8, data, w)}
\description{Scatmat Tour Plotting
Plots the Scatmat Tour in tab g8}
\keyword{internal}
\author{Bei Huang\email{beihuang@iastate.edu}, Di Cook \email{dicook@iastate.edu}, and Hadley Wickham \email{hadley@rice.edu}}
