\name{.interface_stars}
\alias{.interface_stars}
\title{Stars Tour Plotting...}
\usage{.interface_stars(g4, data, w)}
\description{Stars Tour Plotting
Plots the Stars Tour in tab g4}
\keyword{internal}
\author{Bei Huang\email{beihuang@iastate.edu}, Di Cook \email{dicook@iastate.edu}, and Hadley Wickham \email{hadley@rice.edu}}
