\name{.interface_stereo}
\alias{.interface_stereo}
\title{Stereo Tour Plotting...}
\usage{.interface_stereo(g6, data, w)}
\description{Stereo Tour Plotting
Plots the Stereo Tour in tab g6}
\keyword{internal}
\author{Bei Huang\email{beihuang@iastate.edu}, Di Cook \email{dicook@iastate.edu}, and Hadley Wickham \email{hadley@rice.edu}}
