\name{.interface_xy}
\alias{.interface_xy}
\title{Scatterplot Tour Plotting...}
\usage{.interface_xy(g1, data, w)}
\description{Scatterplot Tour Plotting
Plots the scatterplot Tour in tab g1}
\keyword{internal}
\author{Bei Huang\email{beihuang@iastate.edu}, Di Cook \email{dicook@iastate.edu}, and Hadley Wickham \email{hadley@rice.edu}}
