\name{gui_andrews}
\alias{gui_andrews}
\title{Andrews Tour GUI...}
\usage{gui_andrews(data=flea, ...)}
\description{Andrews Tour GUI                                  
Displays an Andrews Tour GUI}
\details{(Paragraph Description: Explain what it does)}
\author{Bei Huang\email{beihuang@iastate.edu}, Di Cook \email{dicook@iastate.edu}, and Hadley Wickham \email{hadley@rice.edu}}
\keyword{hplot}
\arguments{\item{data}{matrix, or data frame containing numeric columns, defaults to flea dataset}
\item{...}{other arguments passed on to \code{animate} and \code{display_xy}}}
\examples{gui_andrews(flea)}
