\name{gui_density}
\alias{gui_density}
\title{Density Tour GUI...}
\usage{gui_density(data=flea, ...)}
\description{Density Tour GUI                                   
Displays a Density Tour GUI}
\details{(Paragraph Description: Explain what it does)}
\author{Bei Huang\email{beihuang@iastate.edu} and Di Cook \email{dicook@iastate.edu}, and Hadley Wickham \email{hadley@rice.edu}}
\keyword{display_density}
\arguments{\item{data}{matrix, or data frame containing numeric columns, defaults to flea dataset}
\item{...}{other arguments passed on to \code{\link{animate}} and \code{\link{display_xy}}}}
\examples{\dontrun{gui_density(flea)}}
