\name{gui_faces}
\alias{gui_faces}
\title{Face Tour GUI...}
\usage{gui_faces(data=flea, ...)}
\description{Face Tour GUI                                   
Displays an Face Tour GUI}
\details{(Paragraph Description: Explain what it does)}
\author{Bei Huang\email{beihuang@iastate.edu} and Di Cook \email{dicook@iastate.edu}}
\keyword{display_faces}
\arguments{\item{data}{matrix, or data frame containing numeric columns, defaults to flea dataset}
\item{...}{other arguments passed on to \code{\link{animate}} and \code{\link{display_xy}}}}
\examples{\dontrun{gui_faces(flea)}}
