\name{gui_image}
\alias{gui_image}
\title{Image Tour GUI...}
\usage{gui_image(data=ozone, ...)}
\description{Image Tour GUI                                   
Displays an Image Tour GUI}
\details{(Paragraph Description: Explain what it does)}
\author{Bei Huang\email{beihuang@iastate.edu} and Di Cook \email{dicook@iastate.edu}}
\keyword{display_image}
\arguments{\item{data}{matrix, or data frame containing numeric columns, defaults to flea dataset}
\item{...}{other arguments passed on to \code{\link{animate}} and \code{\link{display_xy}}}}
\examples{\dontrun{gui_image(ozone)}}
