\name{gui_pcp}
\alias{gui_pcp}
\title{PCP Tour GUI...}
\usage{gui_pcp(data=flea, ...)}
\description{PCP Tour GUI                                   
Displays an PCP Tour GUI}
\details{(Paragraph Description: Explain what it does)}
\author{Bei Huang\email{beihuang@iastate.edu} and Di Cook \email{dicook@iastate.edu}}
\keyword{display_pcp}
\arguments{\item{data}{matrix, or data frame containing numeric columns, defaults to flea dataset}
\item{...}{other arguments passed on to \code{\link{animate}} and \code{\link{display_xy}}}}
\examples{gui_pcp(flea)}
