\name{gui_scatmat}
\alias{gui_scatmat}
\title{Scatmat Tour GUI...}
\usage{gui_scatmat(data=flea, ...)}
\description{Scatmat Tour GUI                                   
Displays an Scatmat Tour GUI}
\details{GUI allows users to control the Scatter matrix tour by simply moving and clicking their mouses.}
\author{Bei Huang\email{beihuang@iastate.edu}, Di Cook \email{dicook@iastate.edu}, and Hadley Wickham \email{hadley@rice.edu}}
\keyword{display_scatmat}
\arguments{\item{data}{matrix, or data frame containing numeric columns, defaults to flea dataset}
\item{...}{other arguments passed on to \code{\link{animate}} and \code{\link{display_xy}}}}
\examples{\dontrun{gui_scatmat(flea)}}
