\name{gui_stars}
\alias{gui_stars}
\title{Stars Tour GUI...}
\usage{gui_stars(data=flea, ...)}
\description{Stars Tour GUI                                   
Displays an Stars Tour GUI}
\details{GUI allows users to control the stars tour by simply moving and clicking their mouses.}
\author{Bei Huang\email{beihuang@iastate.edu}, Di Cook \email{dicook@iastate.edu}, and Hadley Wickham \email{hadley@rice.edu}}
\keyword{display_stars}
\arguments{\item{data}{matrix, or data frame containing numeric columns, defaults to flea dataset}
\item{...}{other arguments passed on to \code{\link{animate}} and \code{\link{display_xy}}}}
\examples{\dontrun{gui_stars(flea)}}
