\name{gui_stereo}
\alias{gui_stereo}
\title{Stereo Tour GUI...}
\usage{gui_stereo(data=flea, ...)}
\description{Stereo Tour GUI                                   
Displays an Stereo Tour GUI}
\details{(Paragraph Description: Explain what it does)}
\author{Bei Huang\email{beihuang@iastate.edu} and Di Cook \email{dicook@iastate.edu}}
\keyword{display_stereo}
\arguments{\item{data}{matrix, or data frame containing numeric columns, defaults to flea dataset}
\item{...}{other arguments passed on to \code{\link{animate}} and \code{\link{display_xy}}}}
\examples{gui_stereo(flea)}
