\name{.create_1d_tour}
\alias{.create_1d_tour}
\title{Density Tour Plotting...}
\usage{.create_1d_tour(data, var_selected, method_selected, center_selected,
    tour_type, aps)}
\description{Density Tour Plotting
Plots the Density Tour}
\keyword{internal}
\author{Bei Huang\email{beihuang@iastate.edu} and Di Cook \email{dicook@iastate.edu}}
