\name{.create_face_tour}
\alias{.create_face_tour}
\title{Faces Tour Plotting...}
\usage{.create_face_tour(data, var_selected, VarIndex, dim_selected,
    tour_type, aps)}
\description{Faces Tour Plotting
Plots the Faces Tour}
\keyword{internal}
\author{Bei Huang\email{beihuang@iastate.edu} and Di Cook \email{dicook@iastate.edu}}
