\name{.create_pcp_tour}
\alias{.create_pcp_tour}
\title{PCP Tour Plotting...}
\usage{.create_pcp_tour(data, var_selected, cat_selected, dim_selected,
    tour_type, guided_type, lambda, aps)}
\description{PCP Tour Plotting
Plots the PCP Tour}
\keyword{internal}
\author{Bei Huang\email{beihuang@iastate.edu}, Di Cook \email{dicook@iastate.edu}, and Hadley Wickham \email{hadley@rice.edu}}
