\name{gui_scatmat}
\alias{gui_scatmat}
\title{Scatmat Tour GUI...}
\usage{gui_scatmat(data=flea, ...)}
\description{Scatmat Tour GUI                                   
Displays an Scatmat Tour GUI}
\details{(Paragraph Description: Explain what it does)}
\author{Bei Huang\email{beihuang@iastate.edu} and Di Cook \email{dicook@iastate.edu}}
\keyword{display_scatmat}
\arguments{\item{data}{matrix, or data frame containing numeric columns, defaults to flea dataset}
\item{...}{other arguments passed on to \code{\link{animate}} and \code{\link{display_xy}}}}
\examples{\dontrun{gui_scatmat(flea)}}
