\name{gui_stars}
\alias{gui_stars}
\title{Stars Tour GUI...}
\usage{gui_stars(data=flea, ...)}
\description{Stars Tour GUI                                   
Displays an Stars Tour GUI}
\details{(Paragraph Description: Explain what it does)}
\author{Bei Huang\email{beihuang@iastate.edu} and Di Cook \email{dicook@iastate.edu}}
\keyword{display_stars}
\arguments{\item{data}{matrix, or data frame containing numeric columns, defaults to flea dataset}
\item{...}{other arguments passed on to \code{\link{animate}} and \code{\link{display_xy}}}}
\examples{gui_stars(flea)}
