\name{gui_tour}
\alias{gui_tour}
\title{Tour GUI...}
\usage{gui_tour(data=flea, ...)}
\description{Tour GUI                                   
Displays all types of Tour GUI in different tabs}
\details{Combines all of the tour gui's into one, putting a separate tab for each.}
\author{Bei Huang\email{beihuang@iastate.edu}, Di Cook \email{dicook@iastate.edu}, and Hadley Wickham \email{hadley@rice.edu}}
\keyword{display}
\arguments{\item{data}{matrix, or data frame containing numeric columns, defaults to flea dataset}
\item{...}{other arguments passed on to \code{\link{animate}} and \code{\link{display_xy}}}}
\examples{\dontrun{gui_tour(flea)}}
