\name{gui_xy}
\alias{gui_xy}
\title{Scatterplot Tour GUI...}
\usage{gui_xy(data=flea, ...)}
\description{Scatterplot Tour GUI
A graphical user interface enabling interactive control of scatterplot tour.}
\details{(Paragraph Description: Explain what it does)}
\author{Bei Huang\email{beihuang@iastate.edu}, Di Cook \email{dicook@iastate.edu}, and Hadley Wickham \email{hadley@rice.edu}}
\keyword{hplot}
\arguments{\item{data}{matrix, or data frame containing numeric columns, defaults to flea dataset}
\item{...}{other arguments passed on to \code{animate} and \code{display_xy}}}
\examples{gui_xy(flea)}
